Move zauth to core
